requirements.txt
# Copy Trading Platform - Python Dependencies
# Install: pip install -r requirements.txt

# ============================================
# CORE FRAMEWORK
# ============================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
uvloop==0.19.0

# ============================================
# DATABASE
# ============================================
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# ============================================
# DATA VALIDATION
# ============================================
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# ============================================
# AUTHENTICATION & SECURITY
# ============================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# ============================================
# HTTP CLIENT (with connection pooling)
# ============================================
httpx[http2]==0.25.2

# ============================================
# CACHING & MESSAGING
# ============================================
redis==5.0.1
nats-py==2.6.0

# ============================================
# CONFIGURATION
# ============================================
python-dotenv==1.0.0

# ============================================
# MONITORING
# ============================================
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# ============================================
# TESTING (optional, for development)
# ============================================
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# ============================================
# CODE QUALITY (optional, for development)
# ============================================
black==23.12.0
flake8==6.1.0
mypy==1.7.1